+++ dexterous_developer_internal/src/bevy_support/cold.rs
@@ -70,6 +70,11 @@ impl<'a> ReloadableApp for ReloadableAppContents<'a> {
         self
     }
 
+    fn add_event<T: ReplacableEvent>(&mut self) -> &mut Self {
+        self.0.add_event::<T>();
+        self
+    }
+
     fn add_state<S: super::ReplacableState>(&mut self) -> &mut Self {
         self.0.add_state::<S>();
         self